+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Hero Agatha
+- "Triggerer" keyword to Generic skills
+
+### Fixed
+
+- Skill are no longer only matched if they are the first 
+
 ## [0.0.14] - 2023-10-14
 
 ### Added